requirements.txt
PyQt5
scikit-learn
numpy
Pillow
opencv-python
